# src/main/resources/rest/v3.api-docs.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "TrainingTool REST API",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:8080/trainingtool",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "authBasic": []
    }
  ],
  "paths": {
    "/REST/v1/users": {
      "get": {
        "tags": [
          "ws-controller"
        ],
        "summary": "Return all users (without their registrations)",
        "operationId": "getUsers",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionResourceUserDTO"
                }
              }
            }
          }
        }
      }
    },
    "/REST/v1/users/{username}": {
      "get": {
        "tags": [
          "ws-controller"
        ],
        "summary": "Return a specific user",
        "operationId": "getUser",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserDTO"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/REST/v1/regs": {
      "get": {
        "tags": [
          "ws-controller"
        ],
        "summary": "Return registration statistical data",
        "operationId": "getRegs",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionResourceStatDTO"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CollectionResourceUserDTO": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserDTO"
            }
          },
          "total": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "UserDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "username": {
            "type": "string"
          },
          "pnr": {
            "type": "string"
          },
          "roles": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "MEMBER",
                "MANAGER",
                "HR",
                "TRAINING",
                "ADMIN"
              ]
            }
          },
          "sector": {
            "type": "string"
          },
          "function": {
            "type": "string"
          },
          "lastname": {
            "type": "string"
          },
          "firstname": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "managername": {
            "type": "string"
          },
          "subContractor": {
            "type": "boolean"
          }
        }
      },
      "CollectionResourceStatDTO": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StatDTO"
            }
          },
          "total": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "StatDTO": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "registrations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StatRegDTO"
            }
          }
        }
      },
      "StatRegDTO": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          },
          "count": {
            "type": "integer",
            "format": "int32"
          }
        }
      }
    },
    "securitySchemes": {
      "authBasic": {
        "type": "http",
        "description": "Basic Authorization",
        "scheme": "basic"
      }
    }
  }
}